api/requirements.txt
Django==4.2
django_filter
djangorestframework
dj-rest-auth
djangorestframework-simplejwt
django-cors-headers
psycopg
boto3